requirements.txt
Django==1.6.6
MySQL-python>=1.2.3
boto>=2.32.1
djangorestframework>=2.4.2
rsa==3.1.4
South>=1.0
django-aggregate-if>=0.4
psycopg2
pinax-theme-bootstrap==5.5.1
django-user-accounts==1.0c9
django-bootstrap-form==3.1
django-braces==1.4.0
django-datatables-view==1.12
pycrypto==2.6.1

# Macx: $ brew install libtiff libjpeg webp littlecms
# Ubuntu: $ sudo apt-get install libtiff4-dev libjpeg8-dev zlib1g-dev libfreetype6-dev liblcms1-dev libwebp-dev tcl8.5-dev tk8.5-dev
Pillow==2.6.1